refactor(http): set retry handler for all queries
