+++ src/pages/_app.tsx
@@ -3,6 +3,7 @@ import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import type { AppProps } from "next/app";
 import { useRouter } from "next/router";
 
+import { HttpError, HttpErrorCode } from "@/shared/api/types";
 import { setupZodErrorMap } from "@/shared/utils/zod-errors";
 
 import { useAuthGuard } from "@/utils/use-auth-guard";
@@ -12,7 +13,22 @@ import "../css/base.css";
 
 setupZodErrorMap();
 
-const queryClient = new QueryClient();
+const retryHandler = (_: number, error: unknown): boolean => {
+  if (!(error instanceof HttpError)) return true;
+  if (error.code === HttpErrorCode.ServerError) return true;
+  if (error.code === HttpErrorCode.ClientError) return false;
+  return true;
+};
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: retryHandler,
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+    },
+  },
+});
 
 const MyApp: React.FunctionComponent<AppProps> = ({ Component, pageProps }) => {
   const router = useRouter();
